+++ dorsly/src/components/App.jsx
@@ -1,10 +1,13 @@
 import React from 'react'
 import Header from './header'
+import Footer from './footer'
+
 
 export default function App() {
   return (
-
-    <Header/>
-
+    <>
+      <Header/>
+      <Footer/>
+    </>
   )
 }
